+++ packages/openapi-parser/src/configuration.ts
@@ -25,6 +25,8 @@ export const ERRORS = {
   OPENAPI_VERSION_NOT_SUPPORTED:
     'Cannot find supported Swagger/OpenAPI version in specification, version must be a string.',
   INVALID_REFERENCE: 'Can’t resolve URI: %s',
+  EXTERNAL_REFERENCE_NOT_SUPPORTED:
+    'External references are not supported yet: %s',
 } as const
 
 export type VALIDATOR_ERROR = keyof typeof ERRORS
