+++ packages/openapi-parser/src/lib/Validator/Validator.ts
@@ -9,7 +9,7 @@ import {
 } from '../../configuration'
 import type { AnyObject, Filesystem, ValidateResult } from '../../types'
 import { details as getOpenApiVersion } from '../../utils'
-import { resolveRefs } from '../resolve'
+import { resolveReferences } from '../resolve'
 import { checkReferences } from '../resolve/checkReferences'
 import { transformErrors } from './transformErrors'
 
@@ -37,7 +37,7 @@ export class Validator {
   public specification: AnyObject
 
   resolveReferences(filesystem?: Filesystem) {
-    return resolveRefs(
+    return resolveReferences(
       filesystem.find((file) => file.isEntrypoint === true).specification,
     )
   }
