+++ packages/openapi-parser/src/lib/resolve/index.ts
@@ -1,1 +1,1 @@
-export * from './resolveRefs'
+export * from './resolveReferences'
