--- packages/openapi-parser/src/lib/resolve/replaceRefs.ts
@@ -1,6 +1,0 @@
-import { resolve } from './resolve'
-
-export function replaceRefs(tree) {
-  // Don’t mutate the original tree
-  return resolve(structuredClone(tree), true)
-}
