--- packages/openapi-parser/src/lib/resolve/resolveRefs.ts
@@ -1,5 +1,0 @@
-import { replaceRefs } from './replaceRefs'
-
-export function resolveRefs(specification: any) {
-  return replaceRefs(specification)
-}
